Bug fixes.
git-svn-id: https://unisim.org/svn/devel@10913 ea533459-3c0a-0410-a024-a90cf4b01ff5

+++ unisim_lib/unisim/component/cxx/processor/tesla/exec.hh
@@ -204,7 +204,7 @@ template<class CONFIG>
 VectorRegister<CONFIG> Cos(VectorRegister<CONFIG> const & a);
 
 inline uint32_t FPToFX(float f);
-inline float FXToFP(uint32_t f);
+inline double FXToFP(uint32_t f);
 
 template<class CONFIG>
 VectorRegister<CONFIG> FSMin(VectorRegister<CONFIG> const & a, VectorRegister<CONFIG> const & b);
